+++ crates/objc2/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Added `mutability::IsMainThreadOnly`.
+
+
 ## 0.4.1 - 2023-07-31
 
 ### Added